chore: 🐋 test 파일을 __test__ 폴더로 이동 및 훅 테스트 파일 분리
