+++ src/components/atoms/Button/Button/__test__/Button.test.tsx
@@ -1,74 +1,9 @@
-import { render, renderHook, screen } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import type { ButtonAlterAs, ButtonDefault, ButtonProps } from ".";
-import Button from "./Button";
-import useButton from "./useButton";
+import Button from "..";
 
 describe("Button", () => {
-  describe("useButton 함수 테스트", () => {
-    it("renderAs가 정의되지 않았을 때, renderAs의 기본값이 button이어야 합니다", () => {
-      const props = {};
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current.renderAs).toBe("button");
-    });
-
-    it("renderAs가 a로 정의되었을 때, role이 button이어야 합니다", () => {
-      const props = { renderAs: "a" as const };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current.role).toBe("button");
-    });
-
-    it("renderAs가 'a'이고 disabled가 false일 때, tabIndex가 0이 되어야합니다.", () => {
-      const props = { renderAs: "a" as const, disabled: false };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current.tabIndex).toBe(0);
-    });
-
-    it("disabled가 true일 때, aria-disabled가 true가 되어야합니다.", () => {
-      const props = { disabled: true };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current["aria-disabled"]).toBe(true);
-    });
-
-    it("label이 정의되었을 때, aria-label이 해당 label값을 가져야합니다.", () => {
-      const label = "Test Label";
-      const props = { label };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current["aria-label"]).toBe(label);
-    });
-
-    it("renderAs가 'a'이고 disabled가 false일 때, onClick이 원래의 onClick함수를 가져야합니다.", () => {
-      const onClick = jest.fn();
-      const props = { renderAs: "a" as const, disabled: false, onClick };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current.onClick).toBe(onClick);
-    });
-
-    it("renderAs가 'a'이고 disabled가 true일 때, onClick이 undefined가 되어야합니다.", () => {
-      const props = { renderAs: "a" as const, disabled: true, onClick: jest.fn() };
-      const { result } = renderHook(() => useButton(props));
-
-      expect(result.current.onClick).toBeUndefined();
-    });
-
-    it("otherProps가 정의되었을 때, 결과 객체에 otherProps가 포함되어야합니다.", () => {
-      const otherProps = { dataTestId: "test" };
-      const props = { ...otherProps };
-      const { result } = renderHook(() =>
-        useButton(props as ButtonProps<ButtonDefault, ButtonAlterAs>)
-      );
-
-      expect(result.current).toHaveProperty("dataTestId", "test");
-    });
-  });
-
   describe("button 태그로 렌더링", () => {
     it("에러 없이 렌더링되어야 합니다.", () => {
       render(<Button>Button</Button>);
